NRL-758 Ephemeral/shared for dynamodb with module
WARNING: This is a WIP commit.

A pointer table module has been created by copy-pasting from the account
wide infrastructure, but changing name to 'document-pointer' to avoid
making changes in the code.

That means that the database works fine for ephemeral environments but
won't work for shared infrastructure as the table name for the shared
infrastructure is 'pointers-table' instead of 'document-pointer'.

The next commit should fix that by using 'pointers-table' name for
everything and making the code changes.