NRL-758 Use permission store module
WARNING: This is a WIP commit

The permissions-store-bucket has been copy pasted from account wide
infrastructure but the name of the bucket has been changed to match
the one the code is expecting.

This means that this commit works well for ephemeral environments but
persistent environments will break as the shared bucket is named
differently.